Añadido ejemplo practico de retutilizacion de componentes Jetpack Compose
