+++ ReusableComponentsDemo/gradle.properties
