+++ ReusableComponentsDemo/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Jan 02 19:53:27 CET 2025
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.7-bin.zip
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
