--- Reutilizaciondecomponentes/.idea/.name
@@ -1,1 +1,0 @@
-Reutilizacion de componentes
