--- Reutilizaciondecomponentes/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Thu Jan 02 18:01:51 CET 2025
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.7-bin.zip
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
